correct programs and rename programs
